1-12-2022, Opensearch clusters reconciler was improved

+++ controllers/clusters/opensearch_controller.go
@@ -33,6 +33,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
+	clusterresourcesv1alpha1 "github.com/instaclustr/operator/apis/clusterresources/v1alpha1"
 	clustersv1alpha1 "github.com/instaclustr/operator/apis/clusters/v1alpha1"
 	"github.com/instaclustr/operator/pkg/instaclustr"
 	convertorsv1 "github.com/instaclustr/operator/pkg/instaclustr/api/v1/convertors"
@@ -68,45 +69,27 @@ func (r *OpenSearchReconciler) Reconcile(ctx context.Context, req ctrl.Request) 
 	err := r.Client.Get(ctx, req.NamespacedName, openSearch)
 	if err != nil {
 		if k8serrors.IsNotFound(err) {
-			return reconcile.Result{}, nil
+			logger.Info("OpenSearch cluster resource is not found",
+				"Request", req,
+			)
+
+			return models.ReconcileResult, nil
 		}
 
-		logger.Error(err, "unable to fetch OpenSearch cluster")
-		return reconcile.Result{}, err
+		logger.Error(err, "unable to fetch OpenSearch cluster resource",
+			"Request", req,
+		)
+
+		return models.ReconcileResult, nil
 	}
 
-	openSearchAnnotations := openSearch.Annotations
-	switch openSearchAnnotations[models.ResourceStateAnnotation] {
+	switch openSearch.Annotations[models.ResourceStateAnnotation] {
 	case models.CreatingEvent:
-		err = r.HandleCreateCluster(openSearch, &logger, &ctx)
-		if err != nil {
-			logger.Error(err, "cannot create OpenSearch cluster",
-				"Cluster name", openSearch.Spec.Name,
-			)
-			return reconcile.Result{}, err
-		}
-
-		return reconcile.Result{Requeue: true}, nil
+		return r.HandleCreateCluster(ctx, openSearch, logger), nil
 	case models.UpdatingEvent:
-		reconcileResult, err := r.HandleUpdateCluster(&ctx, openSearch, &logger)
-		if err != nil {
-			if errors.Is(err, instaclustr.ClusterNotRunning) {
-				logger.Info("OpenSearch cluster is not ready to update",
-					"Cluster name", openSearch.Spec.Name,
-					"Cluster status", openSearch.Status.ClusterStatus,
-					"Reason", err)
-				return *reconcileResult, nil
-			}
-
-			logger.Error(err, "cannot update OpenSearch cluster",
-				"Cluster name", openSearch.Spec.Name,
-			)
-			return reconcile.Result{}, err
-		}
-
-		return *reconcileResult, nil
+		return r.HandleUpdateCluster(ctx, openSearch, logger), nil
 	case models.DeletingEvent:
-		return *r.HandleDeleteCluster(&ctx, openSearch, &logger), nil
+		return r.HandleDeleteCluster(ctx, openSearch, logger), nil
 	case models.GenericEvent:
 		logger.Info("Opensearch resource generic event",
 			"Cluster manifest", openSearch.Spec,
@@ -125,56 +108,76 @@ func (r *OpenSearchReconciler) Reconcile(ctx context.Context, req ctrl.Request) 
 }
 
 func (r *OpenSearchReconciler) HandleCreateCluster(
+	ctx context.Context,
 	openSearch *clustersv1alpha1.OpenSearch,
-	logger *logr.Logger,
-	ctx *context.Context,
-) error {
-	logger.Info(
-		"Creating OpenSearch cluster",
-		"Cluster name", openSearch.Spec.Name,
-		"Data centres", openSearch.Spec.DataCentres,
-	)
-
-	openSearchSpec := convertorsv1.OpenSearchToInstAPI(&openSearch.Spec)
-
-	id, err := r.API.CreateCluster(instaclustr.ClustersCreationEndpoint, openSearchSpec)
-	if err != nil {
-		logger.Error(
-			err, "cannot create OpenSearch cluster",
+	logger logr.Logger,
+) reconcile.Result {
+	if openSearch.Status.ID == "" {
+		logger.Info(
+			"Creating OpenSearch cluster",
 			"Cluster name", openSearch.Spec.Name,
-			"Cluster manifest", openSearch.Spec,
+			"Data centres", openSearch.Spec.DataCentres,
 		)
-		return err
+
+		openSearchSpec := convertorsv1.OpenSearchToInstAPI(&openSearch.Spec)
+
+		id, err := r.API.CreateCluster(instaclustr.ClustersCreationEndpoint, openSearchSpec)
+		if err != nil {
+			logger.Error(err, "cannot create OpenSearch cluster",
+				"Cluster name", openSearch.Spec.Name,
+				"Cluster spec", openSearch.Spec,
+			)
+
+			return models.ReconcileRequeue
+		}
+
+		openSearch.Annotations[models.ResourceStateAnnotation] = models.CreatedEvent
+		openSearch.Annotations[models.DeletionConfirmed] = models.False
+		openSearch.Finalizers = append(openSearch.Finalizers, models.DeletionFinalizer)
+
+		err = r.patchClusterMetadata(ctx, openSearch, logger)
+		if err != nil {
+			logger.Error(err, "cannot patch OpenSearch cluster resource metadata",
+				"Cluster name", openSearch.Spec.Name,
+				"Cluster metadata", openSearch.ObjectMeta,
+			)
+
+			return models.ReconcileRequeue
+		}
+
+		patch := openSearch.NewPatch()
+		openSearch.Status.ID = id
+		err = r.Status().Patch(ctx, openSearch, patch)
+		if err != nil {
+			logger.Error(err, "cannot update OpenSearch cluster resource status",
+				"Cluster name", openSearch.Spec.Name,
+				"Cluster status", openSearch.Status,
+			)
+
+			return models.ReconcileRequeue
+		}
+
+		if len(openSearch.Spec.TwoFactorDelete) != 0 || openSearch.Spec.Description != "" {
+			err = r.updateDescriptionAndTwoFactorDelete(openSearch)
+			if err != nil {
+				logger.Error(err, "cannot update description and twoFactorDelete",
+					"Cluster name", openSearch.Spec.Name,
+					"TwoFactorDelete", openSearch.Spec.TwoFactorDelete,
+					"Description", openSearch.Spec.Description,
+				)
+
+				return models.ReconcileRequeue
+			}
+		}
 	}
 
-	openSearch.Annotations[models.ResourceStateAnnotation] = models.UpdatingEvent
-	openSearch.Finalizers = append(openSearch.Finalizers, models.DeletionFinalizer)
-
-	err = r.patchClusterMetadata(ctx, openSearch, logger)
+	err := r.startClusterStatusJob(openSearch)
 	if err != nil {
-		logger.Error(err, "cannot patch OpenSearch cluster",
-			"Cluster name", openSearch.Spec.Name,
-			"Cluster metadata", openSearch.ObjectMeta,
+		logger.Error(err, "cannot start OpenSearch cluster status job",
+			"Cluster ID", openSearch.Status.ID,
 		)
-		return err
-	}
 
-	patch := openSearch.NewPatch()
-	openSearch.Status.ID = id
-	err = r.Status().Patch(*ctx, openSearch, patch)
-	if err != nil {
-		logger.Error(err, "cannot update OpenSearch cluster status",
-			"Cluster name", openSearch.Spec.Name,
-			"Cluster status", openSearch.Status,
-		)
-		return err
-	}
-
-	err = r.startClusterStatusJob(openSearch)
-	if err != nil {
-		logger.Error(err, "cannot start cluster status job",
-			"OpenSearch cluster ID", openSearch.Status.ID)
-		return err
+		return models.ReconcileRequeue
 	}
 
 	logger.Info(
@@ -186,70 +189,62 @@ func (r *OpenSearchReconciler) HandleCreateCluster(
 		"Namespace", openSearch.Namespace,
 	)
 
-	return nil
+	return models.ReconcileResult
 }
 
 func (r *OpenSearchReconciler) HandleUpdateCluster(
-	ctx *context.Context,
+	ctx context.Context,
 	openSearch *clustersv1alpha1.OpenSearch,
-	logger *logr.Logger,
-) (*reconcile.Result, error) {
+	logger logr.Logger,
+) reconcile.Result {
 	openSearchInstClusterStatus, err := r.API.GetClusterStatus(openSearch.Status.ID, instaclustr.ClustersEndpointV1)
 	if err != nil {
-		logger.Error(
-			err, "cannot get OpenSearch cluster status from the Instaclustr API",
+		logger.Error(err, "cannot get OpenSearch cluster status from the Instaclustr API",
 			"Cluster name", openSearch.Spec.Name,
 			"Cluster ID", openSearch.Status.ID,
 		)
 
-		return &reconcile.Result{}, err
-	}
-
-	if openSearchInstClusterStatus.Status != models.RunningStatus {
-		return &reconcile.Result{
-			Requeue:      true,
-			RequeueAfter: models.Requeue60,
-		}, instaclustr.ClusterNotRunning
-	}
-
-	err = r.reconcileDataCentresNodeSize(openSearchInstClusterStatus, openSearch, logger)
-	if errors.Is(err, instaclustr.StatusPreconditionFailed) {
-		logger.Info("OpenSearch cluster is not ready to resize",
-			"Cluster name", openSearch.Spec.Name,
-			"Cluster status", openSearchInstClusterStatus.Status,
-			"Reason", err,
-		)
-		return &reconcile.Result{
-			Requeue:      true,
-			RequeueAfter: models.Requeue60,
-		}, nil
-	}
-	if errors.Is(err, instaclustr.HasActiveResizeOperation) {
-		logger.Info("Cluster is not ready to resize",
-			"Cluster name", openSearch.Spec.Name,
-			"New data centre manifest", openSearch.Spec.DataCentres[0],
-			"Reason", err,
-		)
-		return &reconcile.Result{
-			Requeue:      true,
-			RequeueAfter: models.Requeue60,
-		}, nil
-	}
-	if err != nil {
-		logger.Error(err, "cannot reconcile data centres node size",
-			"Cluster name", openSearch.Spec.Name,
-			"Cluster status", openSearchInstClusterStatus.Status,
-		)
-		return &reconcile.Result{}, err
+		return models.ReconcileRequeue
 	}
 
 	err = r.updateDescriptionAndTwoFactorDelete(openSearch)
 	if err != nil {
 		logger.Error(err, "cannot update description and twoFactorDelete",
 			"Cluster name", openSearch.Spec.Name,
+			"TwoFactorDelete", openSearch.Spec.TwoFactorDelete,
+			"Description", openSearch.Spec.Description,
+		)
+
+		return models.ReconcileRequeue
+	}
+
+	if openSearchInstClusterStatus.Status != models.RunningStatus {
+		logger.Info("OpenSearch cluster is not ready to update",
+			"Cluster Name", openSearch.Spec.Name,
+			"Reason", instaclustr.ClusterNotRunning,
+		)
+
+		return models.ReconcileRequeue
+	}
+
+	err = r.reconcileDataCentresNodeSize(openSearchInstClusterStatus, openSearch, logger)
+	if errors.Is(err, instaclustr.StatusPreconditionFailed) || errors.Is(err, instaclustr.HasActiveResizeOperation) {
+		logger.Info("OpenSearch cluster is not ready to resize",
+			"Cluster name", openSearch.Spec.Name,
+			"Cluster status", openSearchInstClusterStatus.Status,
+			"New data centre spec", openSearch.Spec.DataCentres[0],
+			"Reason", err,
+		)
+
+		return models.ReconcileRequeue
+	}
+	if err != nil {
+		logger.Error(err, "cannot reconcile data centres node size",
+			"Cluster name", openSearch.Spec.Name,
 			"Cluster status", openSearchInstClusterStatus.Status,
 		)
-		return &reconcile.Result{}, err
+
+		return models.ReconcileRequeue
 	}
 
 	openSearch.Annotations[models.ResourceStateAnnotation] = models.UpdatedEvent
@@ -260,7 +255,8 @@ func (r *OpenSearchReconciler) HandleUpdateCluster(
 			"Cluster name", openSearch.Spec.Name,
 			"Cluster ID", openSearch.Status.ID,
 		)
-		return &reconcile.Result{}, err
+
+		return models.ReconcileRequeue
 	}
 
 	err = r.patchClusterMetadata(ctx, openSearch, logger)
@@ -269,18 +265,20 @@ func (r *OpenSearchReconciler) HandleUpdateCluster(
 			"Cluster name", openSearch.Spec.Name,
 			"Cluster metadata", openSearch.ObjectMeta,
 		)
-		return &reconcile.Result{}, err
+
+		return models.ReconcileRequeue
 	}
 
 	patch := openSearch.NewPatch()
 	openSearch.Status.ClusterStatus = *openSearchInstClusterStatus
-	err = r.Status().Patch(*ctx, openSearch, patch)
+	err = r.Status().Patch(ctx, openSearch, patch)
 	if err != nil {
 		logger.Error(err, "cannot update OpenSearch cluster status",
 			"Cluster name", openSearch.Spec.Name,
 			"Cluster status", openSearch.Status,
 		)
-		return &reconcile.Result{}, err
+
+		return models.ReconcileRequeue
 	}
 
 	logger.Info("OpenSearch cluster was updated",
@@ -288,14 +286,14 @@ func (r *OpenSearchReconciler) HandleUpdateCluster(
 		"Cluster status", openSearch.Status,
 	)
 
-	return &reconcile.Result{}, nil
+	return models.ReconcileResult
 }
 
 func (r *OpenSearchReconciler) HandleDeleteCluster(
-	ctx *context.Context,
+	ctx context.Context,
 	openSearch *clustersv1alpha1.OpenSearch,
-	logger *logr.Logger,
-) *reconcile.Result {
+	logger logr.Logger,
+) reconcile.Result {
 	if openSearch.DeletionTimestamp == nil {
 		openSearch.Annotations[models.ResourceStateAnnotation] = models.UpdatedEvent
 		err := r.patchClusterMetadata(ctx, openSearch, logger)
@@ -305,13 +303,13 @@ func (r *OpenSearchReconciler) HandleDeleteCluster(
 				"Cluster name", openSearch.Spec.Name,
 				"Cluster ID", openSearch.Status.ID,
 			)
-			return &models.ReconcileRequeue
+			return models.ReconcileRequeue
 		}
 
 		logger.Info("OpenSearch cluster is no longer deleting",
 			"Cluster name", openSearch.Spec.Name,
 		)
-		return &models.ReconcileResult
+		return models.ReconcileResult
 	}
 
 	status, err := r.API.GetClusterStatus(openSearch.Status.ID, instaclustr.ClustersEndpointV1)
@@ -321,10 +319,27 @@ func (r *OpenSearchReconciler) HandleDeleteCluster(
 			"Cluster status", openSearch.Status.Status,
 		)
 
-		return &models.ReconcileRequeue
+		return models.ReconcileRequeue
 	}
 
 	if status != nil {
+		if len(openSearch.Spec.TwoFactorDelete) != 0 && openSearch.Annotations[models.DeletionConfirmed] != models.True {
+			logger.Info("OpenSearch cluster deletion is not confirmed",
+				"Cluster name", openSearch.Spec.Name,
+				"Annotations", openSearch.Annotations,
+			)
+
+			openSearch.Annotations[models.ResourceStateAnnotation] = models.UpdatingEvent
+			err = r.patchClusterMetadata(ctx, openSearch, logger)
+			if err != nil {
+				logger.Error(err, "cannot patch OpenSearch cluster metadata",
+					"Cluster name", openSearch.Spec.Name,
+				)
+
+				return models.ReconcileRequeue
+			}
+		}
+
 		err = r.API.DeleteCluster(openSearch.Status.ID, instaclustr.ClustersEndpointV1)
 		if err != nil {
 			logger.Error(err, "cannot delete OpenSearch cluster",
@@ -332,14 +347,14 @@ func (r *OpenSearchReconciler) HandleDeleteCluster(
 				"Cluster status", openSearch.Status.Status,
 			)
 
-			return &models.ReconcileRequeue
+			return models.ReconcileRequeue
 		}
 
 		logger.Info("OpenSearch cluster is being deleted",
 			"Cluster name", openSearch.Spec.Name,
 			"Cluster status", openSearch.Status.Status,
 		)
-		return &models.ReconcileRequeue
+		return models.ReconcileRequeue
 	}
 
 	r.Scheduler.RemoveJob(openSearch.GetJobID(scheduler.StatusChecker))
@@ -351,7 +366,8 @@ func (r *OpenSearchReconciler) HandleDeleteCluster(
 			"Cluster name", openSearch.Spec.Name,
 			"Cluster ID", openSearch.Status.ID,
 		)
-		return &models.ReconcileRequeue
+
+		return models.ReconcileRequeue
 	}
 
 	logger.Info("OpenSearch cluster was deleted",
@@ -359,7 +375,7 @@ func (r *OpenSearchReconciler) HandleDeleteCluster(
 		"Cluster ID", openSearch.Status.ID,
 	)
 
-	return &models.ReconcileResult
+	return models.ReconcileResult
 }
 
 func (r *OpenSearchReconciler) startClusterStatusJob(cluster *clustersv1alpha1.OpenSearch) error {
@@ -380,7 +396,8 @@ func (r *OpenSearchReconciler) newWatchStatusJob(cluster *clustersv1alpha1.OpenS
 		if err != nil {
 			return err
 		}
-		if cluster.DeletionTimestamp != nil {
+
+		if clusterresourcesv1alpha1.IsClusterBeingDeleted(cluster.DeletionTimestamp, len(cluster.Spec.TwoFactorDelete), cluster.Annotations[models.DeletionConfirmed]) {
 			l.Info("OpenSearch cluster is being deleted. Status check job skipped",
 				"Cluster name", cluster.Spec.Name,
 				"Cluster ID", cluster.Status.ID,
@@ -431,16 +448,18 @@ func (r *OpenSearchReconciler) SetupWithManager(mgr ctrl.Manager) error {
 				return true
 			},
 			UpdateFunc: func(event event.UpdateEvent) bool {
-				if event.ObjectOld.GetGeneration() == event.ObjectNew.GetGeneration() {
-					return false
-				}
-
-				if event.ObjectNew.GetDeletionTimestamp() != nil {
-					event.ObjectNew.GetAnnotations()[models.ResourceStateAnnotation] = models.DeletingEvent
+				newObj := event.ObjectNew.(*clustersv1alpha1.OpenSearch)
+				if newObj.DeletionTimestamp != nil &&
+					(len(newObj.Spec.TwoFactorDelete) == 0 || newObj.Annotations[models.DeletionConfirmed] == models.True) {
+					newObj.Annotations[models.ResourceStateAnnotation] = models.DeletingEvent
 					return true
 				}
 
-				event.ObjectNew.GetAnnotations()[models.ResourceStateAnnotation] = models.UpdatingEvent
+				if event.ObjectOld.GetGeneration() == newObj.Generation {
+					return false
+				}
+
+				newObj.Annotations[models.ResourceStateAnnotation] = models.UpdatingEvent
 				return true
 			},
 			GenericFunc: func(genericEvent event.GenericEvent) bool {
